+++ google/cloud/google_cloud_cpp_common.bzl
@@ -17,6 +17,7 @@
 """Automatically generated source lists for google_cloud_cpp_common - DO NOT EDIT."""
 
 google_cloud_cpp_common_hdrs = [
+    "backoff_policy.h",
     "future.h",
     "future_generic.h",
     "future_void.h",
