+++ google/cloud/spanner/CMakeLists.txt
@@ -38,7 +38,6 @@ include(GoogleCloudCppCommon)
 configure_file(version_info.h.in ${CMAKE_CURRENT_SOURCE_DIR}/version_info.h)
 add_library(
     spanner_client # cmake-format: sort
-    backoff_policy.h
     backup.cc
     backup.h
     batch_dml_result.h
