+++ google/cloud/spanner/internal/connection_impl.cc
@@ -71,9 +71,9 @@ std::unique_ptr<RetryPolicy> DefaultConnectionRetryPolicy() {
 
 std::unique_ptr<BackoffPolicy> DefaultConnectionBackoffPolicy() {
   auto constexpr kBackoffScaling = 2.0;
-  return google::cloud::spanner::ExponentialBackoffPolicy(
-             std::chrono::milliseconds(100), std::chrono::minutes(1),
-             kBackoffScaling)
+  return google::cloud::ExponentialBackoffPolicy(std::chrono::milliseconds(100),
+                                                 std::chrono::minutes(1),
+                                                 kBackoffScaling)
       .clone();
 }
