+++ google/cloud/spanner/internal/retry_loop.h
@@ -15,9 +15,9 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_INTERNAL_RETRY_LOOP_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_INTERNAL_RETRY_LOOP_H
 
-#include "google/cloud/spanner/backoff_policy.h"
 #include "google/cloud/spanner/retry_policy.h"
 #include "google/cloud/spanner/version.h"
+#include "google/cloud/backoff_policy.h"
 #include "google/cloud/internal/invoke_result.h"
 #include "google/cloud/status_or.h"
 #include <grpcpp/grpcpp.h>
