+++ google/cloud/spanner/polling_policy.h
@@ -15,9 +15,9 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_POLLING_POLICY_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_SPANNER_POLLING_POLICY_H
 
-#include "google/cloud/spanner/backoff_policy.h"
 #include "google/cloud/spanner/retry_policy.h"
 #include "google/cloud/spanner/version.h"
+#include "google/cloud/backoff_policy.h"
 #include "google/cloud/status.h"
 
 namespace google {
